+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Corrected HID report descriptor
 * Add max_packet_size() to HidDevice to allow differing report sizes
 * Allows default layer to be set on a Layout externally
+* Add Chording for multiple keys pressed at the same time to equal another key
 
 Breaking changes:
 * Row and Column pins are now a simple array. For the STM32 MCU, you
@@ -20,6 +21,7 @@ Breaking changes:
   example), managing a mouse emulation, or any other ideas you can
   have. As there is a default value for the type parameter, the update
   should be transparent.
+* Layers don't sum anymore, the last pressed layer action set the layer.
 * Rename MeidaCoffee in MediaCoffee to fix typo.
 
 # v0.1.1